lib/backend/requirements.txt
fastapi
uvicorn
python-pptx
nbconvert
nbformat
pytest
pytest-asyncio
python-multipart
html2image
